#363 Added support for Long property conversion into EntityResourceUtils. Removed unnecessary narrowing of Long versions to Integer-typed versions.

+++ platform-pojo-bl/src/main/java/ua/com/fielden/platform/entity/AbstractFunctionalEntityForCollectionModification.java
@@ -40,15 +40,15 @@ public abstract class AbstractFunctionalEntityForCollectionModification<K extend
     
     @IsProperty
     @Title(value = "Surrogate Version", desc = "Surrogate Version")
-    private Integer surrogateVersion;
+    private Long surrogateVersion;
 
     @Observable
-    public AbstractFunctionalEntityForCollectionModification<K> setSurrogateVersion(final Integer surrogateVersion) {
+    public AbstractFunctionalEntityForCollectionModification<K> setSurrogateVersion(final Long surrogateVersion) {
         this.surrogateVersion = surrogateVersion;
         return this;
     }
 
-    public Integer getSurrogateVersion() {
+    public Long getSurrogateVersion() {
         return surrogateVersion;
     }
